Switch to cats-based implementation
This makes it easier to abstract both over the Json values and the patch
type, making it possible to add new patch format and diff algorithms
more easily.
It is also safer to use in case of exceptions.

This new implementation focuses on the core feature, no (un)marshalling
is included in the core library, which only manipulates Json values for
the underlying Json provider. This results in a much cleaner interface
with less responsabilities.